Contain Python AST emission behind compiler adapter
